code_examples/cinemabot/requirements.txt
aiogram
aiohttp
pypika
bs4
Pillow
